Make this stack standalone by not requiring (or using) a default VPC.
This adds a VPC, and introduces some interesting complexity in that the prefix list ID for Instance Conect and the Security Group ID for Global Accelerator can't be known beforehand (the SG ID may not even be present unless an Accelerator has already been created in the region).  So, a custom resource lambda is employed to grab those values. It works, but needs some armor (if the SG is missing, wait and retry a few times).

The EC2 instance script works now, but is hard-coded w/t ports, which can be remedied with more Fn::ForEach -- more remaining work.